update and create export request, store, wishlist of branch section

+++ src/components/branch/exportRequest/CreateItem.js
@@ -1,32 +1,13 @@
 import React,{ useState } from 'react'
 import axios from 'axios'
 import { useParams, useHistory } from 'react-router-dom'
+import CreateItemModal from './CreateItemModal'
 
 const CreateItem = (props) => {
   const history = useHistory()
   const { requestId } = useParams()
   const [status,setStatus] = useState(props.status)
-  const [code,setCode] = useState('')
-  const [quantity,setQuantity] = useState(0)
-  const [error,setError] = useState('')
-  const [success,setSuccess] = useState('')
-
-  const submitHandler = async e => {
-    e.preventDefault()
-    try{
-      const add = await axios.post(`http://localhost:1000/api/exportRequest/${requestId}/item`,{ code, quantity })
-      if(add){
-        setError('')
-        console.log(add.data);
-        setSuccess("Item added!")
-        setCode('')
-        setQuantity(0)
-      }
-    }catch(err){
-      setSuccess('')
-      setError(err.response.data)
-    }
-  }
+  const [show,setShow] = useState(false)
 
   const sentRequest = async e => {
     const getSend = await axios.put(`http://localhost:1000/api/exportRequest/${requestId}/send`)
@@ -43,81 +24,53 @@ const CreateItem = (props) => {
   }
 
   const deleteRequest = async e => {
-    const getDelete = await axios.delete(`http://localhost:1000/api/exportRequest/${requestId}`)
-    if(getDelete){
-      history.push('/exportRequest')
+    const sure = window.confirm('Are you sure you want to delete?')
+    if(sure){
+      const getDelete = await axios.delete(`http://localhost:1000/api/exportRequest/${requestId}`)
+      if(getDelete){
+        history.push('/exportRequest')
+      }
     }
   }
 
   return(
-    <div className="py-3">
+    <div className="mb-4 create-item">
       <div className="d-flex">
         {status === 'create' && <button
           type="button"
-          className="btn btn-sm btn-light"
-          data-toggle="collapse"
-          data-target="#createItem"
+          className="btn rounded-circle btn-light mx-2"
+          title="Create New Item"
+          onClick={e => setShow(true)}
         >
-          <i className="fa fa-plus text-success"></i> Add New Item
+          <i className="fa fa-plus text-success"></i>
         </button>}
         {status === 'create' && props.itemLength > 0 &&  <button
           type="button"
-          className="btn btn-sm btn-light mx-3"
+          className="btn rounded-circle btn-light mx-2"
+          title="Sent Export Request"
           onClick={sentRequest}
         >
-          <i className="fa fa-paper-plane text-success"></i> Sent Request
+          <i className="fa fa-paper-plane text-info"></i>
         </button>}
         {status === 'pending' && <button
           type="button"
-          className="btn btn-sm btn-light mx-3"
+          className="btn rounded-circle btn-light mx-2"
           onClick={cancelRequest}
+          title="Cancel Request"
         >
-          <i className="fa fa-window-close text-warning"></i> Cancel Request
+          <i className="fa fa-window-close text-warning"></i>
         </button>}
         {(status === 'create' || status === 'pending') && <button
           type="button"
-          className="btn btn-sm btn-light mx-3"
+          className="btn rounded-circle btn-light mx-3"
           onClick={deleteRequest}
+          title="Delete Request"
         >
-          <i className="fa fa-trash text-danger"></i> Delete Request
+          <i className="fa fa-trash text-danger"></i>
         </button>}
       </div>
       <div className="collapse mt-4" id="createItem">
-        <form onSubmit={submitHandler} className="row light-border">
-          {success && <div className="col-12 mb-3">
-            <p className="text-center text-success mb-0">{success}</p>
-          </div>}
-          <div className="form-group col-md-6 col-lg-4">
-            <label htmlFor="item_code">Item Code</label>
-            <input
-              type="text"
-              id="item_code"
-              required
-              value={code}
-              className="form-control"
-              onChange={e => setCode(e.target.value)}
-              placeholder="Code"
-            />
-            {error && error.code && <small className="text-danger">*{error.code}</small>}
-          </div>
-          <div className="form-group col-md-6 col-lg-4">
-            <label htmlFor="item_quantity">Item Quantity</label>
-            <input
-              type="number"
-              id="item_quantity"
-              required
-              value={quantity}
-              className="form-control"
-              onChange={e => setQuantity(e.target.value)}
-              placeholder="Quantity"
-              min="0"
-            />
-            {error && error.quantity && <small className="text-danger">*{error.quantity}</small>}
-          </div>
-          <div className="col-12">
-            <button type="submit" className="btn btn-sm btn-success">Create</button>
-          </div>
-        </form>
+        <CreateItemModal requestId={requestId} show={show} setShow={setShow} />
       </div>
     </div>
   )
